requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
openai>=1.0.0,<2.0.0
flask-cors==4.0.0
httpx>=0.24.0,<0.28.0
discord.py>=2.3.0
anthropic>=0.7.0
requests>=2.31.0
